refactor: Split derived.test.js to separate files.
- Deleted the unreadable `derived.test.js` (4668 lines) file and split each suite to its file.
- Moved all specs in widget folder to __tests__ folder

This refactor aims to improve maintainability and ensure that the table widget's derived properties are thoroughly tested.
