+++ app/client/src/widgets/TableWidgetV2/widget/__tests__/derived.test/fixture.js
@@ -1,0 +1,191 @@
+export const samplePrimaryColumns = {
+  step: {
+    index: 0,
+    width: 150,
+    id: "step",
+    originalId: "step",
+    alias: "step",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "text",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isCellVisible: true,
+    isCellEditable: false,
+    isDerived: false,
+    label: "step",
+    computedValue: ["#1", "#2", "#3"],
+    labelColor: "#FFFFFF",
+    cellBackground: "",
+    textColor: "",
+    fontStyle: "",
+  },
+  task: {
+    index: 1,
+    width: 150,
+    id: "task",
+    originalId: "task",
+    alias: "task",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "text",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isCellVisible: true,
+    isCellEditable: false,
+    isDerived: false,
+    label: "task",
+    computedValue: [
+      "Drop a table",
+      "Create a query fetch_users with the Mock DB",
+      "Bind the query using => fetch_users.data",
+    ],
+    labelColor: "#FFFFFF",
+    cellBackground: "",
+    textColor: "",
+    fontStyle: "",
+  },
+  status: {
+    index: 2,
+    width: 150,
+    id: "status",
+    originalId: "status",
+    alias: "status",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "text",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isCellVisible: true,
+    isCellEditable: false,
+    isDerived: false,
+    label: "status",
+    computedValue: ["✅", "--", "--"],
+    labelColor: "#FFFFFF",
+    cellBackground: "",
+    textColor: "",
+    fontStyle: "",
+  },
+  action: {
+    index: 3,
+    width: 150,
+    id: "action",
+    originalId: "action",
+    alias: "action",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "button",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isCellVisible: true,
+    isCellEditable: false,
+    isDisabled: false,
+    isDerived: false,
+    label: "action",
+    onClick:
+      "{{currentRow.step === '#1' ? showAlert('Done', 'success') : currentRow.step === '#2' ? navigateTo('https://docs.appsmith.com/core-concepts/connecting-to-data-sources/querying-a-database',undefined,'NEW_WINDOW') : navigateTo('https://docs.appsmith.com/core-concepts/displaying-data-read/display-data-tables',undefined,'NEW_WINDOW')}}",
+    computedValue: ["", "", ""],
+    labelColor: "#FFFFFF",
+    buttonColor: ["#553DE9", "#553DE9", "#553DE9"],
+    borderRadius: ["0.375rem", "0.375rem", "0.375rem"],
+    boxShadow: ["none", "none", "none"],
+    buttonLabel: ["Action", "Action", "Action"],
+    buttonVariant: "PRIMARY",
+  },
+  customColumn1: {
+    allowCellWrapping: false,
+    index: 4,
+    width: 150,
+    originalId: "customColumn1",
+    id: "customColumn1",
+    alias: "customColumn1",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "menuButton",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isDisabled: false,
+    isCellEditable: false,
+    isEditable: false,
+    isCellVisible: true,
+    isDerived: true,
+    label: "menu",
+    isSaveVisible: true,
+    isDiscardVisible: true,
+    computedValue: "",
+    buttonStyle: "rgb(3, 179, 101)",
+    labelColor: "#FFFFFF",
+    menuColor: ["#553DE9", "#553DE9", "#553DE9"],
+    borderRadius: ["0.375rem", "0.375rem", "0.375rem"],
+    boxShadow: ["none", "none", "none"],
+    customAlias: "",
+    menuButtonLabel: ["Open Menu", "Open Menu", "Open Menu"],
+    menuVariant: "PRIMARY",
+    menuButtoniconName: "",
+  },
+  customColumn2: {
+    allowCellWrapping: false,
+    index: 5,
+    width: 150,
+    originalId: "customColumn2",
+    id: "customColumn2",
+    alias: "customColumn2",
+    horizontalAlignment: "LEFT",
+    verticalAlignment: "CENTER",
+    columnType: "iconButton",
+    textSize: "0.875rem",
+    enableFilter: true,
+    enableSort: true,
+    isVisible: true,
+    isDisabled: false,
+    isCellEditable: false,
+    isEditable: false,
+    isCellVisible: true,
+    isDerived: true,
+    label: "icon",
+    isSaveVisible: true,
+    isDiscardVisible: true,
+    computedValue: "",
+    buttonStyle: "rgb(3, 179, 101)",
+    labelColor: "#FFFFFF",
+    buttonColor: ["#553DE9", "#553DE9", "#553DE9"],
+    borderRadius: ["0.375rem", "0.375rem", "0.375rem"],
+    boxShadow: ["none", "none", "none"],
+    customAlias: "",
+    buttonVariant: "PRIMARY",
+    iconName: "add",
+  },
+};
+export const sampleProcessedTableData = [
+  {
+    step: "#1",
+    task: "Drop a table",
+    status: "✅",
+    action: "",
+    __originalIndex__: 0,
+  },
+  {
+    step: "#2",
+    task: "Create a query fetch_users with the Mock DB",
+    status: "--",
+    action: "",
+    __originalIndex__: 1,
+  },
+  {
+    step: "#3",
+    task: "Bind the query using => fetch_users.data",
+    status: "--",
+    action: "",
+    __originalIndex__: 2,
+  },
+];
